model classes basically completed & create test data in test case

+++ src/main/java/com/nullteam/ragpicker/repository/WxUserRepository.java
@@ -3,5 +3,5 @@ package com.nullteam.ragpicker.repository;
 import com.nullteam.ragpicker.model.WxUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WxUserRepository extends JpaRepository<WxUser, Long> {
+public interface WxUserRepository extends JpaRepository<WxUser, Integer> {
 }
